[Bronze II] Title: 개구리 1, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/19563. 개구리 1/개구리 1.cc
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main() {
-    long long a, b, c;
+    int a, b, c;
 
     cin >> a >> b >> c;
